feat: revenue --not-finished

+++ database/migrations/2025_01_12_145440_create_ship_bays_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->foreign('room_id')->references('id')->on('rooms')->onDelete('cascade');
             $table->string('port');
             $table->json('arena');
-            $table->unsignedBigInteger('revenue')->default(0);
-            $table->unsignedBigInteger('penalty')->default(0);
+            $table->decimal('revenue', 15, 2)->default(0);
+            $table->decimal('penalty', 15, 2)->default(0);
             $table->timestamps();
         });
     }
